Sync side of changes to fix the initial load of the react app.
This includes several changes to loggin in sync/index.tsx,
It changes _processing events to _syncing and _created events,
which allows _created to be triggered before any errors that
prevent the database to start syncing (which was the case for CSIRO
Geochem, as the DB wasn't public. I will keep this inaccessable as a
regression tests for this fix. Basically, CSIRO db errored when it
tried to start syncing, and that error didn't count towards the total
count of DB's created, so it never completed.)