Remove uiSpec stuff from projectNav, made staging a constant timer.
This staging timer is now started whenever a formikContext is available.
(Perhaps this should instead be run when formikProps is available?)
From this formikContext it can get values out of the form.
(But in practice only returns empty strings for some reason TBD.)

Also, I thought I had removed uiSpec from projectNav before, but I ended
up needing to do it again because of the 2 places where it can occur.

Also, datamodel was amended to allow non-string form values.