Fix a blank loading screen on first load of blank DB
This was caused by projects_known, and therefore a render, being emitted
before createdProjects was filled. After projects_known is emitted,
the GUI doesn't ever try to refresh the list of projects.

To fix this, we can listen for project_local event to add projects, but
this commit adds a 'projects_created' event for all projects_locals
emitted as well as projects_known. Also, createdProjects setter in
process_project needed to be moved so that createdProjects for a project
is set before project_local emitted.