Comment out App test
This was the example test from CRA, and is now more of an integration or
e2e test rather than a unit test. Moving to integration_tests and
commenting out for now, we may try to get it working again (or replace
it with other tests), but for now drop it as its not helpful.