Sync on/off overhaul to fix sync behaviour after toggling
Previously, switching sync on/off would turn off pushes.
And it would also remove event listeners.

This commit fixes that by adding a SyncHandler, and pull/push options,
to LocalDBRemote data, and moving out the ensure_synced_db logic to
setLocalConnection, so it can also be used by the sync toggle.

This means that the initial db sync and further syncs all use the same
logic to set pull/push options and setup event emitters.

Also added is a thing that got caught due to these changes: SyncHandler
didn't have a way to cancel its timeouts when the DB it was listening to
is destroyed.

Signed-off-by: Aidan Farrell <aidan.farrell@mq.edu.au>