'Variants' of revisions added to FAIMS.
This is a big commit because it's a change to the data model.
The previous commit only added variants to the project UI spec, but did
not do anything with the variants.

This commit adds variants to the schema for the data DB,
(They could have gone in records instead of revisions, but this would
stop us from in the future supporting changing variants of existing records)

When a Record is created (record-create.tsx), the Record form is not
immediately shown. Instead the user is provided a list of buttons to
choose which type of record they want to create. (This is untested)
This is passed to the RecordForm component instead of a Revision ID that
is passed to RecordForm if it's displaying an existing record.

When RecordForm views an existing record, it knows the variant by
fetching the existing record and using the 'variant' property that this
commit adds to the data model.

I removed start_view, replacing it with the first view in the selected
variant. (grep for '[0]' to put it back)

I also found I needed to clean up the Record Form state/props handling:
variant name, current view name and revision id are all things that can
be passed into the RecordForm prop, but might not be, or might change.
These are therefore all stored in the component state and are updated
from the props in formChanged(). Coming from these, It turns out that
the staging area was a bit iffy with regard to which things it considered
optional, so that's been a bit fixed up too.

Something I came accross while implementing: There is a significant
repeated code to do with converting a Promise/Error to React components,
I did a history.goBack() when errors occur, which might not be the best.

Signed-off-by: Aidan Farrell <aidan.farrell@mq.edu.au>

Fix one of the broken tests caused by modified ProjectUIModel

Signed-off-by: Aidan Farrell <aidan.farrell@mq.edu.au>