Get rid of SyncHandler, do everything based on PouchDB .changes
This commit moves us towards being able to properly handle changes in
the directory & listings, as well as changes in active db (without
AUTOACTIVATE, which is a debug thing).

We get rid of some events that should be instead done using .changes
& awaiting getProjectDB/getDataDB functions.

This also means that any changes to the local DB that DON'T Come from
a remote (i.e. new tabs, or different react elements that are siloed
off from each other, don't affect each others rendering) should be
properly updated, because _paused events are no longer to be used
outside initialization, the rest of the codebase will use .changes now.

Also, state.ts is significantly changed to be a more user visible
and usable way to know if everything's initialzed, and access to the
databases.

Signed-off-by: Aidan Farrell <aidan.farrell@mq.edu.au>