e2e: disable reloading newly created record because hrid is not shown on the list yet
There is no way to know which record we just created because the list is showing database id, not hrid. Database id is only briefly shown in the current URL upon clicking "SAVE AND NEW", but then gets overriden by the new record. I was told hrid will one day be shown in the list, so we'll revisit then.

Signed-off-by: riniangreani <rini.angreani@csiro.au>