Fix syncing with attachment handling
There was some confusion between which direction push and pull operated
due to the ordering of databases being passed to sync. By setting the
order such that push is app -> couch and pull couch -> app, the
filtering is now being applied correctly.

There was also a typo for the filter name, which resulted in 404s.

Signed-off-by: James Tocknell <aragilar@gmail.com>