Add service worker
The docs weren't clear on the need to use a custom template to get the
service worker for offline use. This adds the additional files that are
needed from https://github.com/cra-template/pwa/, as well as the workbox
dependencies.

See https://create-react-app.dev/docs/making-a-progressive-web-app/ for
more details about configuring it.

Signed-off-by: James Tocknell <aragilar@gmail.com>